MG349
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@936 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Stylization/GDUtils.cpp
@@ -207,13 +207,13 @@ void rs_gdImageMultiPolygon(gdImagePtr im, int* cntrs, int nCntrs, gdPointPtr p,
 void rs_gdImageCircleForBrush(gdImagePtr im, int x, int y, int rad, RS_Color& color)
 {
     int gdc = ConvertColor(im, color);
-    float rad2 = rad*rad;
+    float rad2 = (float)rad*rad;
 
     for (int j = -rad, k = y+j; j <= rad; j++, k++)
     {
         float j_offset = j + 0.5f;
 
-        float hlen = sqrt((float)(rad2 - j_offset*j_offset));
+        float hlen = sqrt(rad2 - j_offset*j_offset);
 
         int solidwid = (int)hlen;
